+++ Source/R.DirectX.6.0.3DS/Renderer.cxx
@@ -2140,7 +2140,7 @@ namespace RendererModule
     }
 
     // 0x60004050
-    void SelectRendererFogAlphas(const u32* input, u8* output)
+    void SelectRendererFogAlphas(const u8* input, u8* output)
     {
         if (input == NULL) { return; }
